+++ src/quick-lint-js/i18n/translation-table-generated.h
@@ -20,8 +20,8 @@ namespace quick_lint_js {
 using namespace std::literals::string_view_literals;
 
 constexpr std::uint32_t translation_table_locale_count = 5;
-constexpr std::uint16_t translation_table_mapping_table_size = 441;
-constexpr std::size_t translation_table_string_table_size = 76255;
+constexpr std::uint16_t translation_table_mapping_table_size = 442;
+constexpr std::size_t translation_table_string_table_size = 76303;
 constexpr std::size_t translation_table_locale_table_size = 35;
 
 QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
@@ -78,6 +78,7 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "'if' statement"sv,
           "'in' disallowed in C-style for loop initializer"sv,
           "'infer' is only allowed between 'extends' and '?' in conditional types"sv,
+          "'out in' is not allowed; write 'in out' instead"sv,
           "'private' is not allowed in JavaScript"sv,
           "'protected' is not allowed in JavaScript"sv,
           "'public' is not allowed in JavaScript"sv,